# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skills Latvia 2022 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Displays user list",
                "description": "Displays user list",
                "operationId": "UserList",
                "requestBody": {
                    "description": "Page number",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "page": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new user",
                "description": "Create a new user",
                "operationId": "UserStore",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Displays user by id",
                "description": "Displays user by id",
                "operationId": "UserShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Object ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update an existing user",
                "description": "Update an existing user",
                "operationId": "UserUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Object ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user by id",
                "description": "Delete user by id",
                "operationId": "UserDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Object ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/NullResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "NullResource": {},
...
            "UserStoreRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user1@mail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "PassWord12345"
                    },
                    "name": {
                        "type": "string",
                        "example": "Gibril"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Pavlovic"
                    },
                    "role": {
                        "description": "Role of the user. Allowed role types: admin, courier, client",
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user1@mail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "PassWord12345"
                    },
                    "name": {
                        "type": "string",
                        "example": "Gibril"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Pavlovic"
                    },
                    "role": {
                        "description": "Role of the user. Allowed role types: admin, courier, client",
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "vladislavs"
                    },
                    "surname": {
                        "type": "string",
                        "example": "kuznecovs"
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@dgtcloud.lv"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-04-20T12:00:43.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2022-04-20T12:00:43.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}